[flue] Extract Dependabot review helpers into lib/dependabot-review.ts
